+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### :wrench: Changed
 
 - **Coverage phase**: Generate values for empty schemas.
+- **Coverage phase**: Generate values for `not` keyword and `false` JSON Schemas.
 
 ## [4.0.25](https://github.com/schemathesis/schemathesis/compare/v4.0.24...v4.0.25) - 2025-08-10